+++ examples/append_ndjson/main.go
@@ -26,8 +26,8 @@ import (
 	"log"
 	"time"
 
-	scopedb "github.com/scopedb/scopedb-sdk/go"
-	"github.com/scopedb/scopedb-sdk/go/examples/internal/exampleutil"
+	scopedb "github.com/scopedb/goscopedb"
+	"github.com/scopedb/goscopedb/examples/internal/exampleutil"
 )
 
 func main() {