+++ examples/catalog/main.go
@@ -22,8 +22,8 @@ import (
 	"fmt"
 	"log"
 
-	scopedb "github.com/scopedb/scopedb-sdk/go"
-	"github.com/scopedb/scopedb-sdk/go/examples/internal/exampleutil"
+	scopedb "github.com/scopedb/goscopedb"
+	"github.com/scopedb/goscopedb/examples/internal/exampleutil"
 )
 
 func main() {