+++ examples/ingest_transform/main.go
@@ -25,8 +25,8 @@ import (
 	"log"
 	"time"
 
-	scopedb "github.com/scopedb/scopedb-sdk/go"
-	"github.com/scopedb/scopedb-sdk/go/examples/internal/exampleutil"
+	scopedb "github.com/scopedb/goscopedb"
+	"github.com/scopedb/goscopedb/examples/internal/exampleutil"
 )
 
 func main() {