+++ itcases/ingest_stream_test.go
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"github.com/gkampitakis/go-snaps/snaps"
-	scopedb "github.com/scopedb/scopedb-sdk/go"
+	scopedb "github.com/scopedb/goscopedb"
 	"github.com/stretchr/testify/require"
 )