+++ itcases/table_test.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 	"testing"
 
-	scopedb "github.com/scopedb/scopedb-sdk/go"
+	scopedb "github.com/scopedb/goscopedb"
 	"github.com/stretchr/testify/require"
 )